+++ docs/changelog.md
@@ -8,6 +8,12 @@
 
 See the [GitHub "features" project](https://github.com/peter88213/novelyst/projects/1).
 
+### v0.10.0 Beta release
+
+- Add scene date/time/duration display.
+
+Based on PyWriter v5.6.1
+
 ### v0.8.0 Beta release
 
 - Add Character Bio and Goals entry.